- oops. Missed change of `hudDraw()` `zDelta` from int to double while implementing ae2d2c7948279ba5a0ac5ea4eb5582db53cb9535.

+++ source/blood/src/hudsprites.cpp
@@ -95,7 +95,7 @@ static void viewBurnTime(int gScale)
 }
 
 
-void hudDraw(PLAYER *gView, int nSectnum, int defaultHoriz, double bobx, double boby, int zDelta, int basepal)
+void hudDraw(PLAYER *gView, int nSectnum, int defaultHoriz, double bobx, double boby, double zDelta, int basepal)
 {
 	if (gViewPos == 0)
 	{
